requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fire==0.6.0
fonttools==4.54.1
fsspec==2024.9.0
grpcio==1.66.2
humanfriendly==10.0
idna==3.10
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.3.0
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
megadetector==5.0.19
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.48
protobuf==3.20.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
Pyqtree==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics-yolov5==0.1.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
